Fixed warnings reported by gcc 4.x.
git-svn-id: http://svn.boost.org/svn/boost/trunk@31239 b8fc166d-592f-0410-95f2-cb63ce0dd405

+++ boost/boost/wave/cpplexer/re2clex/cpp_re2c_lexer.hpp
@@ -141,7 +141,7 @@ lexer<IteratorT, PositionT>::get()
     unsigned int actline = scanner.line;
     token_id id = token_id(scan(&scanner));
     
-    switch (id) {
+    switch (static_cast<unsigned int>(id)) {
     case T_IDENTIFIER:
     // test identifier characters for validity (throws if invalid chars found)
         value = string_type((char const *)scanner.tok,
